Move template into one file
This is a big change from how it was done previously.
Now there is one template that holds how the target .hpp file will look
This makes it easier to extend it based on future needs.
It also enables the template to change to match other projects
that may want a different setup to the header.